requirements.txt
black
diffusers[torch]
flake8
numpy
Pillow
pycodestyle
safetensors
transformers
torch
